Make authoring actions work as a widget in authoring react
