+++ package.json
@@ -97,6 +97,7 @@
     "moment": "2.30.1",
     "moment-timezone": "0.5.41",
     "ng-file-upload": "12.2.13",
+    "node-polyfill-webpack-plugin": "^4.1.0",
     "owl.carousel": "2.2.0",
     "patch-package": "^7.0.0",
     "popper-max-size-modifier": "0.2.0",
@@ -124,15 +125,14 @@
     "sass-loader": "^10.5.2",
     "shallow-equal": "^3.1.0",
     "shortid": "2.2.8",
-    "superdesk-ui-framework": "6.1.1",
+    "superdesk-ui-framework": "^6.1.2",
     "ts-loader": "^8.4.0",
     "typescript": "~4.9.5",
     "uuid": "8.3.1",
     "web-animations-js": "^2.3.2",
     "webpack": "^5.95.0",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^5.0.4",
-    "node-polyfill-webpack-plugin": "^4.1.0"
+    "webpack-dev-server": "^5.0.4"
   },
   "devDependencies": {
     "@superdesk/build-tools": "file:build-tools",
