+++ scripts/apps/authoring-react/widget-layout-component.tsx
@@ -7,14 +7,21 @@ import {IAuthoringWidgetLayoutProps} from 'superdesk-api';
  */
 export class AuthoringWidgetLayoutComponent extends React.PureComponent<IAuthoringWidgetLayoutProps> {
     render() {
-        const {header, body, footer} = this.props;
+        const {header, body, footer, bodyPadding = 'medium'} = this.props;
+
+        // Map bodyPadding values to PanelContentBlock padding values
+        const paddingMap: Record<'none' | 'small' | 'medium', '0' | '1-5' | '3'> = {
+            none: '0',
+            small: '1-5',
+            medium: '3',
+        };
 
         return (
             <Layout.Panel side="right" open={true} size="x-small" background={this.props.background ?? 'light'}>
                 {header && <React.Fragment>{header}</React.Fragment>}
 
                 <Layout.PanelContent>
-                    <Layout.PanelContentBlock>
+                    <Layout.PanelContentBlock flex={bodyPadding === 'none'} padding={paddingMap[bodyPadding]}>
                         {body}
                     </Layout.PanelContentBlock>
                 </Layout.PanelContent>
