+++ scripts/core/interactive-article-actions-panel/actions/send-correction-tab.tsx
@@ -1,18 +1,8 @@
 import React from 'react';
 import {IArticle} from 'superdesk-api';
-import {Button, DateTimePicker, ToggleBox} from 'superdesk-ui-framework/react';
-import {gettext, toIsoStringWithoutTimezoneOffset} from 'core/utils';
 import {PanelContent} from '../panel/panel-content';
 import {PanelFooter} from '../panel/panel-footer';
-import ng from 'core/services/ng';
-import {cloneDeep} from 'lodash';
-import {notify} from 'core/notify/notify';
-import {appConfig} from 'appConfig';
-import {getLocaleForDatePicker} from 'core/helpers/ui-framework';
-import {isValid} from 'date-fns';
-import {TimeZonePicker} from 'core/ui/components/time-zone-picker';
-import {TZDate} from '@sourcefabric/date-fns-tz';
-import {ignoreTimezone} from '../subcomponents/publishing-date-options';
+import {SendCorrectionAction} from './send-correction-action';
 
 interface IProps {
     item: IArticle;
@@ -22,126 +12,31 @@ interface IProps {
     onDataChange(changes: IArticle): void;
 }
 
-interface IState {
-    embargo: Date | null;
-    timeZone: string | null;
-}
-
-export class SendCorrectionTab extends React.Component<IProps, IState> {
-    constructor(props: IProps) {
-        super(props);
-
-        this.state = {
-            embargo: props.item.embargo != null ? ignoreTimezone(props.item.embargo) : null,
-            timeZone: props.item.schedule_settings?.time_zone ?? null,
-        };
-
-        this.doSendCorrection = this.doSendCorrection.bind(this);
-    }
-
-    doSendCorrection(): void {
-        this.props.handleUnsavedChanges()
-            .then((item) => {
-                // Cloning to prevent objects from being modified by angular
-                ng.get('authoring').publish(
-                    cloneDeep(item),
-                    cloneDeep(item),
-                    'correct',
-                ).then(() => {
-                    ng.get('authoringWorkspace').close();
-                    notify.success('Correction sent');
-                });
-            })
-            .catch(() => {
-                // cancelled by user
-            });
-    }
-
+/**
+ * Send Correction Tab - composes the SendCorrectionAction into the panel layout.
+ */
+export class SendCorrectionTab extends React.Component<IProps> {
     render() {
-        const {markupV2} = this.props;
-        const {embargo, timeZone} = this.state;
+        const {item, markupV2, closePublishView, handleUnsavedChanges, onDataChange} = this.props;
 
         return (
-            <React.Fragment>
-                <PanelContent markupV2={markupV2}>
-                    <ToggleBox variant="simple" title={gettext('Embargo')} initiallyOpen>
-                        <DateTimePicker
-                            value={embargo}
-                            valueType="date"
-                            locale={{
-                                type: 'full',
-                                payload: getLocaleForDatePicker(),
-                            }}
-                            dateFormat={appConfig.view.dateformat}
-                            onChange={(val) => {
-                                const isValidDate = isValid(val);
-                                const isDateBeingReset = val === null;
-
-                                if (isValidDate || isDateBeingReset) {
-                                    this.setState(
-                                        {
-                                            embargo: val,
-                                            timeZone: timeZone ?? appConfig.default_timezone,
-                                        },
-                                        () => {
-                                            this.props.onDataChange({
-                                                ...this.props.item,
-                                                embargo: val == null
-                                                    ? null
-                                                    : toIsoStringWithoutTimezoneOffset(new TZDate(val)),
-                                                schedule_settings: {
-                                                    ...this.props.item.schedule_settings,
-                                                    time_zone: this.state.timeZone,
-                                                },
-                                            });
-                                        },
-                                    );
-                                }
-                            }}
-                        />
-                    </ToggleBox>
-
-                    {embargo != null && (
-                        <ToggleBox variant="simple" title={gettext('Time zone')} initiallyOpen>
-                            <TimeZonePicker
-                                value={timeZone}
-                                onChange={(val) => {
-                                    this.setState(
-                                        {timeZone: val},
-                                        () => {
-                                            this.props.onDataChange({
-                                                ...this.props.item,
-                                                schedule_settings: {
-                                                    ...this.props.item.schedule_settings,
-                                                    time_zone: val,
-                                                },
-                                            });
-                                        },
-                                    );
-                                }}
-                            />
-
-                            {timeZone == null && (
-                                <div style={{paddingBlockStart: 5}}>
-                                    {gettext('If not set, the UTC+0 time zone is assumed.')}
-                                </div>
-                            )}
-                        </ToggleBox>
-                    )}
-                </PanelContent>
-
-                <PanelFooter markupV2={markupV2}>
-                    <Button
-                        text={gettext('Send correction')}
-                        onClick={() => {
-                            this.doSendCorrection();
-                        }}
-                        size="large"
-                        type="highlight"
-                        expand
-                    />
-                </PanelFooter>
-            </React.Fragment>
+            <SendCorrectionAction
+                item={item}
+                closePublishView={closePublishView}
+                handleUnsavedChanges={handleUnsavedChanges}
+                onDataChange={onDataChange}
+            >
+                {({body, footer}) => (
+                    <>
+                        <PanelContent markupV2={markupV2}>
+                            {body}
+                        </PanelContent>
+                        <PanelFooter markupV2={markupV2}>
+                            {footer}
+                        </PanelFooter>
+                    </>
+                )}
+            </SendCorrectionAction>
         );
     }
 }
